ostd_test

Struct KtestItemInfo

#[repr(C)]
pub struct KtestItemInfo {
    pub module_path: &'static str,
    pub fn_name: &'static str,
    pub package: &'static str,
    pub source: &'static str,
    pub line: usize,
    pub col: usize,
}

The information of the unit test.

Fields
module_path: &'static str

The path of the module, not including the function name.

It would be separated by ::.

fn_name: &'static str

The name of the unit test function.

package: &'static str

The name of the crate.

source: &'static str

The source file where the test function resides.

line: usize

The line number of the test function in the file.

col: usize

The column number of the test function in the file.